Added collision to the hole so that when the balls hits the final hole, it stop the ball power and any of game movement and tells the user they're go a par or bogie or eagle, depending on how strokes it took them to pot the ball. To do the collision for the hole I created a square that fits in the gold ball final hole and did the collision like I did for gold course egdes. To make it seem like I didn't do this, I simply made the square the same colour as the grass background. I only chose a small square as it would mean the ball would have to go slightly in the hole to fully collide. This stops users from touching the egde of the hole and winning the game.
I also added 4 portals to gold course, which act as random transporters
to the other portals. The collision for these holes is like the collision
final hole. I used the srand random number function to do as this means
that the user does not where they will end, thus hopefully making the game
more interesting for the user. To keep the uer from going into one hole and
then constantly transporting between the four holes, I used a bool that is
set at the beginning of each stroke. Therefore, the user can only go
through one portal per stroke. I chose this because it's hard, and
pointless, to go through two portals in one turn.